Revert CodeSandbox back to Parcel
This partially reverts 4027d2ad81cfe7090083ec09c18b7a8b8a10fb4b.

Apparently CodeSandbox interprets the app as a "Node app" when not
using Parcel, which (currently) prevents people from forking it
when not signed in, and presumably will cause it to make
assumptions that do not hold. Since Parcel only caused problems in
setups we're not using, I've reverted the Webpack migration for
now. We can re-enable it if the problems re-appear and CodeSandbox
support is more stable.